refactor alvl repo and service 1

+++ Aero.Application/DTOs/AccessLevelComponentDto.cs
@@ -6,7 +6,7 @@ namespace Aero.Application.DTOs;
 public sealed record AccessLevelComponentDto(
     short DriverId,
     string Mac,
-    short DoorId,
+    int DoorId,
     short AcrId,
     short TimeZoneId
     );
